Make C++ consistent with Java for pdf417
A bunch of small changes that result in C++ and Java equiv. for pdf417.

- A little bit of java re-reporting in pdf417
- A little bit of java re-reporting/consistency elsewhere

There a few places related to latching that differed between the java
and C++. I looked for a discussion of this on the group and didn't find
any (but may have missed it.) I reverted to the java code which fixed
one of the tests. The comments referenced standards-compliance vs. real
world encoders. That may need to be revisited?

git-svn-id: http://zxing.googlecode.com/svn/trunk@2673 59b500cc-1b3d-0410-9834-0bbf25fbcc57