Simplify conversion RGB -> luminance
Removed testing every pixel for greyness in RGBLuminanceSource, as it
did not appreciably change the speed of converting a greyscale image and
slowed down the conversion of a colour image by a few percent.  Besides,
there was no need to treat every row separately; the total number of
pixels was enough.