removed 'continuous-video' mode
I found it better to not use this mode becouse it sets focus on start and then never tweak it again later. 

It is returned in AutofocusManager as currentFocusMode and maks useAutoFocus == false.

For me it worked better if currentFocusMode is 'auto' than 'continous-video' becouse it start autofocusing by focus() and autoFocusAgainLater() by AutoFocusManager. 

"... The focus position is locked after autoFocus call ...."
source:
https://developer.android.com/reference/android/hardware/Camera.Parameters.html#FOCUS_MODE_CONTINUOUS_VIDEO