Introduce BinaryBitmap parent, AbstractBinaryBitmap
This change should help with custom BinaryBitmap implementations, which were previously awkward.
This way was chosen instead of making the BinaryBitmap into an interface to minimize breakage
for end users, as I assume that there are more people using BinaryBitmap than implementing custom Readers.
Making the BinaryBitmap non-final is also an option, but custom subclasses would be fragile with regards
to future BinaryBitmap changes.

This also splits isRotateSupported into two variants, for 90 and 45 degrees, to make it easier to implement
rotation support properly, given that 45 degree rotation is not used by built-in Readers, but is required
to leverage the rotation support.